Move desktop capture API to Screen object
