+++ src/api/dispatcher_host.cc
@@ -172,9 +172,6 @@ void DispatcherHost::OnAllocateObject(int object_id,
     objects_registry_.AddWithID(new Shortcut(object_id, weak_ptr_factory_.GetWeakPtr(), option), object_id);
   } else if (type == "Screen") {
     objects_registry_.AddWithID(new EventListener(object_id, weak_ptr_factory_.GetWeakPtr(), option), object_id);
-  } else if (type == "DesktopCapture") {
-	VLOG(1) << "Initing DesktopCapture object";
-    objects_registry_.AddWithID(new DesktopCapture(object_id, weak_ptr_factory_.GetWeakPtr(), option), object_id);	
   } else {
     LOG(ERROR) << "Allocate an object of unknown type: " << type;
     objects_registry_.AddWithID(new Base(object_id, weak_ptr_factory_.GetWeakPtr(), option), object_id);
