+++ src/api/screen/desktopcapture.cc
